requirements.txt
aiosqlite==0.18.0
alembic==1.9.3
anyio==3.6.2
apprise==1.2.1
asgi-lifespan==2.0.0
asyncpg==0.27.0
attrs==22.2.0
boto3==1.26.74
botocore==1.29.74
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.8
cryptography==39.0.1
dateparser==1.1.7
docker==6.0.1
fastapi==0.90.0
fsspec==2023.1.0
google-auth==2.16.0
greenlet==2.0.2
griffe==0.25.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==25.3.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
mdurl==0.1.2
mypy-boto3-s3==1.26.62
mypy-boto3-secretsmanager==1.26.49
numpy==1.24.2
oauthlib==3.2.2
orjson==3.8.6
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pendulum==2.1.2
prefect @ file:///opt/prefect/dist/prefect.tar.gz
prefect-aws==0.2.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
pydot==1.4.2
Pygments==2.14.0
pyodbc==4.0.35
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.3
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
sqlalchemy-schemadisplay==1.3
starlette==0.23.0
text-unidecode==1.3
toml==0.10.2
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
websocket-client==1.5.1
websockets==10.4
